pyproject.toml
[tool.poetry]
name = "aptrs"
version = "1.0"
description = "APTRS (Automated Penetration Testing Reporting System) is a Python and Django-based automated reporting tool designed for penetration testers and security organizations. This tool streamlines the report generation process by enabling users to create PDF and Excel reports directly, eliminating the need for manual approaches. Additionally, APTRS offers a systematic way to monitor and manage vulnerabilities within various projects. Keep your penetration testing projects organized and efficient with APTRS."
authors = ["Sourav Kalal <kalalsourav20@gmail.com>"]
license = "APTRS Licence"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2.15"
django-phonenumber-field = "7.0.0"
django-phonenumbers = "1.0.1"
djangorestframework = "3.15.2"
gunicorn = "22.0.0"
pillow = "10.4.0"
pytz = "2022.6"
sqlparse = "0.5.0"
tzdata = "^2024.1"
django-cors-headers = "3.13.0"
djangorestframework-simplejwt = "5.3.1"
colorlog = "6.7.0"
django-filter = "23.5"
python-dotenv = "1.0.0"
weasyprint = "^62.3"
pygal = "3.0.4"
bleach = {extras = ["css"], version = "^6.2.0"}
xlsxwriter = "3.1.9"
whitenoise = "6.6.0"
docxtpl = "0.18.0"
redis = "5.0.1"
psycopg2-binary = "^2.9.9"
jinja2 = "^3.1.4"
django-redis = "^5.4.0"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
boto3 = "^1.34.159"
django-storages = "^1.14.4"
bs4 = "^0.0.2"
requests = "^2.32.3"
[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


